invester pitch/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
email-validator
passlib
bcrypt==3.2.2
python-jose[cryptography]
python-multipart
google-auth
requests
sendgrid==6.11.0
slowapi==0.1.9
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
google-generativeai==0.8.0
fastapi-mail
